Add new CI release workflow (#496)
# Goal
The goal of this PR is to implement the new CI Release Workflow per this
diagram:

https://docs.google.com/drawings/d/1IcAzZbulRcntYQ5ub-zDyIkqJY1NE0iqClhkg2PE5lo

This PR restores status quo in CI, but under the new process. Moving
forward will be implementation of new features, pretty much.

Please note:

1. Release is triggered whenever a tag matching these glob strings is
pushed:
   ```sh
      - "v[0-9]+.[0-9]+.[0-9]+" # ex. v0.29.9
      - "v[0-9]+.[0-9]+.[0-9]+-[0-9]+" # ex. v0.29.9-1
      - "v[0-9]+.[0-9]+.[0-9]+-rc[0-9]+" # ex. v0.29.9-1-rc1
   ```
1. Whatever you annotate the release tag with is the content that will
appear on the GitHub release page. We'll probably address this soon in a
future PR depending on what the team wants to be shown in that section.

Part of #282