Feat/nonce buckets against replay attacks (#511)
# Goal
The goal of this PR is to implement and use a replay attack prevention
technique in two extrinsics that take externally created signatures of
externally provided data:
* `create_sponsored_account_with_delegation`
* `add_provider_to_msa`

The replay attack prevention provides an early exit for both of these
extrinsics.

Related to #327 

# Discussion
Sets up a system of rotating virtual buckets to store signatures, using
the bucket number and signature instead of a nonce. Signatures must
provide an expiration block. This is used to compute the bucket number
it's expected to be in, and find out whether the signature has been
recently submitted.

Signatures are rejected if they are found in the registry or the
associated payload has an expiration too far in the future (>=
MortalityBucketSize * NumberOfBuckets)

Every MortalityBucketSize * NumberOfBuckets blocks, the bucket rotation
starts over, the old signatures are cleared, and the new signature is
added.

To be able to check whether a virtual bucket needs to be cleared, bucket
numbers are mapped to a mortality block.

Signature storage is really a
[StorageDoubleMap](https://paritytech.github.io/substrate/master/frame_support/storage/trait.StorageDoubleMap.html)
where K1 = bucket number and K2 is the signature. The value stored is
the expiration provided in `add_provider_payload` (currently unused).
The bucket expiration block, being the first key in the
`StorageDoubleMap` makes it extremely easy to "clear" a virtual bucket
using
[`drain_prefix`](https://paritytech.github.io/substrate/master/frame_support/storage/trait.IterableStorageDoubleMap.html#tymethod.drain_prefix)
(via `IterableStorageDoubleMap` trait).

# Checklist
- [x] Refactor `EXPIRATION_BLOCK_VALIDITY_GAP`?
- [x] Set up genesis for buckets in mocks
- [x] New Errors, Constants in pallet
- [x] Values in mock
- [x] Values in frequency and frequency-rococo runtimes
- [x] StorageMap for the bucket --> mortality map
- [x] StorageDoubleMap for BucketSignature --> mortality
- [x] Helpers in pallet
- [x] Unit tests for helper fns
- [x] Replay tests passing
- [x] N/A Chain spec updated
- [x] N/A Updated the js/api-augment code if a custom RPC added/changed
- [x] N/A Design doc(s) updated
- [x] Tests added
- [x] Benchmarks added
- [x] Weights updated