feat(schemas): update schema permission storage
Update the schema-permission storage to use a
BoundedBTreeMap instead of an OrderSet.

This update increase performance when writing and
updating schema permissions.

Additionally, each schema permission
stores a field, "revoked_at," that indicates the
block number for which the permission expires.
If set to "None" type, it means that
the permission is not revoked and is valid.

issue-455