Bugs/retire msa dos attack #605 (#611)
# Goal
The goal of this PR is to address the retire_msa section of #605 :
ensure that the extrinsic cannot fail and instead, any potential
failures happen in the SignedExtension.

# Discussion
The main change is to move all db access inside a match that fetches the
MSA, primarily for clarity.

A check for an MSA is already made in the SignedExtension, so that it fails if
the Origin (actually the AccountId) does not have an MSA. It also
verifies that there is only one key, the AccountId, associated with it,
and that it's not a registered provider. This means that we should never
get into the extrinsic unless there is an MSA. We're required to provide
a None option in our match so if we somehow strangely wind up in there,
it logs a warning and exits.

- [ ] `ensure_signed` - this needs to stay in the extrinsic, otherwise there is no AccountId to look up an MSA
- [ ] `try_get_msa_from_account_id` - this has been removed, as it is
identical to another function. We know already we should have an MSA, so
we just fetch it instead of doing a try_get type of function call.
- [ ] `remove_delegator` - similarly, this being in the match it doesn't
happen unless there is an MSA. If there is no delegator - which is
allowed, the MSA may already have removed a provider - then there is no
failure; this is harmless.
- [ ]  `delete_key_for_msa` - also in the Some(msa_id) clause as above.

## Also
I refactored away a function that wasn't really doing anything extra.