chore(msa): update benchmarks
Update on_initialize benchmarks
so that it verifies that the
payload-signature-registry storage
is cleared for bucket zero.

Update the mock runtime Msa config,
MortalityWindowSize, so that it is
the same as the frequency runtime.
This update allows running benchmarks
for test and production have the same
configurations. By not having them the same,
the test benchmarks fail to run.

Update test to pass failing test after updating
runtime config MortalityWindowSize.
Update benchmarks.

issue-668