chore(capacity): add storage
Add storage for staking and Capacity:

`StakingAccountLedger`
Storage for keeping a ledger for staking

``StakingTargetLedger`
Storage to record how many tokens were targeted to an MSA.

`CapacityOf``
Storage for target Capacity usage.

These storage items are based on the Capacity design docs.

Adding storage makes it so staking, unstaking, and
withdraw_unstake features can be worked on in parallel.