feat(tx-payment): add ability to use Capacity
Add the ability to use Capacity for transacting.

This implementation wraps the substrate transaction-payment
pallet, and it is used to toggle between token transactions
and capacity transactions.

When Capacity is used to transact, tipping is
set to zero regardless if a user attempts to tip
with Capacity. Token transaction tipping is
still supported.

issue-1026