Feat/whitelist capacity eligible calls #1013 (#1158)
## Goal
The goal of this PR is to set up a whitelist of transactions that are
eligible to be paid for with Capacity.

Closes #1013 

## Discussion
Once I realized this would work if it were set up like `BaseCallFilter`
it was pretty easy to make a Config in the pallet that only needs to
implement the `Contains` trait. So we simply construct a list of call
signatures to match against for a given runtime. Then we apply the
whitelist in `withdraw_fee`.