[Capacity] integration tests for transacting #889 (#1203)
The goal of this PR is to implement integration tests for transacting
with capacity.
Closes #889

Acceptance Criteria
- [x] When a user does not have enough capacity to pay for the
transaction fee and is NOT eligible to replenish Capacity, it should
error and be dropped from the transaction pool.
- [x] When a user attempts to pay for a non-capacity transaction with
Capacity, it should error and drop the transaction from the transaction
pool.
- [x] When Capacity has withdrawn an event CapacityWithdrawn is emitted.
- [x] When a user is not a registered provider and attempts to pay with
Capacity, it should error with InvalidTransaction::Payment.
- [x] Integration test and manual test. A registered provider with
Capacity but no tokens associated with the key should still be able to
use polkadot UI to submit a capacity transaction.
- [x] Token transactions should not be affected by supporting Capacity
transactions.
1. Should all capacity-eligible transactions be tested for success with
pay_with_capacity()?

- [x] Tests added

- [ ] Chain spec updated
- [ ] Custom RPC OR Runtime API added/changed? Updated js/api-augment.
- [ ] Design doc(s) updated
- [ ] Benchmarks added
- [ ] Weights updated