feat(capacity): add minimum token balance
The goal of this PR is to implement a Minimum Token Balance.
Currently, a user is allowed to stake all of their free balance.
This could be inconvenient when you unstake and are left with 0 balance.
This Minimum Token Balance is a configurable amount with a minimum value that must be greater than the Existential Deposit.

Update weights for PR #1256

Closes #944