Capacity staking min token balance (#1256)
# Goal
The goal of this PR is to implement a Minimum Token Balance. 
Currently, a user is allowed to stake all of their free balance. 
This could be inconvenient when you unstake and are left with 0 balance.
This Minimum Token Balance is a configurable amount with a minimum value
that must be greater than the Existential Deposit.

Closes #944 

# Checklist
- [x] Design doc(s) updated
- [x] Tests added
- [x] Weights updated
# N/A
- [ ] Chain spec updated
- [ ] Custom RPC OR Runtime API added/changed? Updated js/api-augment.
- [ ] Benchmarks added