Update integration test scripts and options (#1259)
# Goal
The goal of this PR is to update the run_integration_tests.sh script to:

1. Handle signals and clean up.
2. Add a command line argument to skip the js build.

Change Explanations:

Makefile:
- Added targets to enable skipping js build.
  - integration-test = no arguments
  - integration-test-only = `-s` argument to skip js build
  - integration-test-load = `load` argument
- integration-test-load-only = `-s load` arguments to skip js build for
load tests

run_integration_tests.sh:
- Modify `cleanup` not to use the kill_freq.sh script. As signals are
asynchronous, `cleanup` does not know how to navigate to the scripts
directory using relative paths. It's a small script, so the commands are
included directly in `cleanup`.
- Fix syntax errors so that `getopts` will correctly parse the `-s`
option.
- Add `if-then` conditional for `SKIP_JS_BUILD`.
- Remove `if-then` conditional which is handled by `cleanup` on
receiving the `EXIT` signal when the script finishes.
- Refactor `SHOULD_KILL` and only populate `PID` when the script starts
the Frequency node.