1401 Build/run Frequency runtime without a relay chain (#1464)
# Goal
The goal of this PR is to allow building and running the _Frequency_
runtime without a relay chain. Previously, it ran but was looking for a
relay chain since it was compiled and configured as a _Cumulus_
parachain. As there is now, no relay chain, blocks need to be sealed
using instant sealing (default) or manual sealing.

# Testing
To build Frequency for use without a relay:

`cargo build --features frequency-no-relay`

In addition to basic testing with Polkadot UI, try upgrading the runtime
using `Sudo` (use `sudoUncheckedWeight`) and the `setCode` extrinsic.
You will need to build a 2nd Frequency build that has the `spec_version`
incremented.

`make upgrade-no-relay` also needs testing...

It would be prudent to test other configurations:

`frequency-rococo-local`
`frequency-rococo-testnet`
`frequency`

Closes #1401

---------

Co-authored-by: Dmitri <4452412+demisx@users.noreply.github.com>