Replace production profile with release profile (#1524)
# Goal
The goal of this PR is to reduce confusion and have faster local builds.

Instead of having three main profiles, this reduces it to just two.

Before: `debug`, `release`, `production`
After: `debug`, `release`

Generally:
- `--profile production` -> `--release` (which is the same as `--profile
release`)
- `--release` -> Removed (default `debug`)
- `--locked` -> Removed except for the release builds

The `release` profile wasn't really being used and is the more standard
"production" build. So we pulled the build configuration for production
into `release`

# Discussion
- Also updated all the make builds to use `debug` except ones that need
`release` (which is default)
- This reduces the rebuild time as much as possible, even when switching
builds.