Switch to the new version of jacogr/polkadot-js-tools (#1594)
# Goal
The goal of this PR is to upgrade the polkadotjs tools for metadata
comparison.

Closes #1230 

This also includes PR #1601 for expediency. So also part of #1362

Ended up needing to run each one in a docker container with a new docker
network. This allowed everything to talk correctly.

# Checklist
- [x] Tested with: make start + docker run --rm --platform=linux/amd64
--pull always jacogr/polkadot-js-tools:0.55.3 metadata
wss://1.rpc.frequency.xyz ws://host.docker.internal:9944
- [x] Tested with test release Running:
https://github.com/LibertyDSNP/frequency/actions/runs/5282806513

---------

Co-authored-by: Dmitri <4452412+demisx@users.noreply.github.com>