Increase Capacity Batch limit from 2 -> 10 (#1633)
# Goal
The goal of this PR is to increase the limit on capacity batches from 2
-> 10

Closes #1628

# Discussion
Why? 2 was set originally to be initially careful with how many were
allowed. With additional testing, it is clear there is no reason to not
increase to at least 10. In think in the end we'll be able to increase
to higher, but going slow.

# Checklist
- [x] Weights updated

---------

Co-authored-by: Frequency CI [bot] <do-not-reply@users.noreply.github.com>