# Goal This PR aims to remove the `rustup install toolchain nightly` commands from the individual GitHub jobs that use the `ci-base-image` that previously did not have the `nightly` toolchain.
Closes #1727

 # Discussion
The changes in the previous PR (#1700) to add the `nightly` toolchain to `ci-base-image` were
not published to DockerHub until the `merge-pr.yml` workflow actions were successfully run.
Therefore the `nightly` toolchain was not available for CI to use in `ci-base-image`.

In that PR, the `nightly` toolchain was installed in the individual jobs where it was needed to pass CI.

Now that the PR has been merged, `ci-base-image` has been updated with both the `stable`
and `nightly` toolchains and the redundant `rustup install` commands can be removed from `verify-pr-commit.yml`.