Feat/split storage #1726 (#1744)
The goal of this PR is to split up storage of boosting and maximized staking accounts,
as well as store retarget history separately, which can store retargeting events for
any type of staking.

Closes #1726

* Lots of rattle, causing several helper functions to be split up -
rather than having a bunch of conditionals due to type differences.
* In the interest of limiting the size of this PR/merge into the branch
I left a lot of TODOs for when reward pool updates on a
stake/boost/unstake is resumed