feat: change instant-seal node docker image to use startup script (#1781)
# Goal
The goal of this PR is to allow the use of interval sealing mode in CI
pipelines using GitHub workflows.

The following considerations were taken into account:
* Since GitHub workflows do not allow overriding the `CMD` of a Docker
image, and overriding `ENTRYPOINT` in GitHub likewise does not allow for
passing arguments to the new entrypoint, the simplest option left is to
use environment variables.
* Since it was deemed undesirable to modify the frequency binary itself
to honor environment variables, the approach taken was to create a new
image, `frequencychain/standalone-node`. The `ENTRYPOINT` of the image
is a shell script instead of the `frequency` binary directly. The shell
script takes into account the following variables:
    * `SEALING_MODE`: `instant`, `interval`, or `manual`
* `SEALING_INTERVAL`: number of seconds between blocks in `interval`
sealing mode
* `CREATE_EMPTY_BLOCKS`: `true` means pass
`--sealing-create-empty-blocks` to allow interval sealing mode to form
empty blocks as in a live chain.

Additional arguments passed to the script are passed through to the
underlying `frequency` command-line, just as the `CMD` parameter in the
old image. Existing use cases of this image should therefore be
unaffected.
* The existing `frequencychain/instant-seal-node` image is now
deprecated and will eventually be discontinued/removed.

Closes #1780 

# Discussion
<!-- List discussion items -->

# Checklist
- [x] Doc(s) updated
- [x] Tested new image for all sealing modes: `instant`, `interval`,
`manual`

---------

Co-authored-by: Wil Wade <wil.wade@unfinished.com>