Chore: remove nightly from cargo doc to fix warnings (#1916)
# Goal
The goal of this PR is remove a reference to `+nightly` and resolve some
warnings when executing `make lint`.

# Changes
- Remove `+nightly` in favor of `RUSTC_BOOTSTRAP=1`, which basically
turns the stable toolchain into a nightly toolchain without having to
install the nightly.
**WARNING:** `RUSTC_BOOTSTRAP` is not recommended for production
binaries. It is used here only for `cargo doc`. `RUSTC_BOOTSTRAP=1` is
designed to be used for bootstraping the compiler and removes some
stable chain safety mechanisms.
- Use `1.76.0` instead of `stable`, as it depends on `ci-base-image`
version at build time.

# How to Test
- Execute `make lint` and observe correct completion with no warnings.

# Checklist
- [ ] Chain spec updated
- [ ] Custom RPC OR Runtime API added/changed? Updated js/api-augment.
- [ ] Design doc(s) updated
- [ ] Tests added
- [ ] Benchmarks added
- [ ] Weights updated

---------

Co-authored-by: Matthew Orris <--help>