Benchmark fixes (#1971)
# Goal
The goal of this PR is to fix an issue with running benchmarks and
update our benchmarks with some tests.

Part of #1853

## Issues

- When I added the Proxy pallet, the copied weight file worked, but the
generated version didn't.
- Then when fixing that, the benchmark runners had issues. So updated
the benchmark workflow by removing it from the commit commenting system
to a manual run. Using that will also automatically trigger verify CI
actions. (GitHub actions only run on Manually triggered actions to
prevent loops)

# Discussion
- The additional changes to the templates is from Polkadot-SDK
- Actual change is small, as most of this are the generated weights code
- Fix issue with the proxy_pallet weights reference
- Update the benchmark running system to not use special Github tokens