Fix Missing runtime APIs for Metadata v15 (#2204)
# Goal
The goal of this PR is to make sure that Metadata v15 has the correct
Runtime API call data

Closes #2202 

# Discussion

Why? I cannot be 100% sure. See the linked issues in #2202 for more
details.

From what I can tell, the way rust macros work with context means that
this macro is stuck being in the same file as the const that uses it is
`pub const VERSION`

# Specifics

- Updated Polkadot js API
- Fixed issue with the runtime APIs and the call in
`e2e/capacity/capacity_rpc.test.ts`. Issue noted in code comment
- Moved the runtime apis back into the runtime lib.rs
- Removed the manual definitions of runtimeApis 🎉 

## Testing

- `make start` && `subxt metadata --version 15 -f json | jq
.[1].V15.apis` (Before this result would have been empty, now you have
the runtime data!)