fix graph keys on testnet attempt 2 (#2232)
# Goal

The muti-block migration did not work as expected. It looks like the
code inside `on_initialize` did not get executed. I'm out of theories
why that would happen but there might be something related to the
feature flags. This is another attempt that replaces feature flag
attempt with the genesis hash.

# Changes
- replaced feature flags with genesis hash checks
- Added tighter checks `on_initialize` (for better readability)

# Discussions
If anyone have a theory why the code inside `on_initialize` didn't get
executed on paseo please contact me.

Closes #2227

# Checklist
- [X] Spec version incremented?