fixes double charge, undo changes to increase balances in e2e tests (#2314)
# Goal
Fixes the double-charge problem and undoes some of the hacks used to
make some e2e tests pass.

Followed the model of ChargeTransactionPayment and created a
`can_withdraw_fee` function in the Nontransferable trait. Then this is
used by `validate` and `withdraw_fee` is only used in `pre_dispatch`.

Following the guidance/direction from Parity, we do not perform the
validation steps twice so `withdraw_fee` doesn't call
`can_withdraw_fee`.

This also required adding a `replenishable_balance` to the
Nontransferable trait, which isn't what I'd prefer but the trait
generics required it, otherwise it was a mismatch of Balance types.

---------

Co-authored-by: Aramik <aramikm@gmail.com>
Co-authored-by: Joe Caputo <joseph.caputo@projectliberty.io>