Split the build benchmark binary and benchmark run steps (#2337)
This updates the way benchmarks are run so that we use less time on the
benchmark runner.

Benchmark runner time is now < 20 minutes. (Total run is much longer,
but this is much easier to use spot instances with)

- Update the release workflow
- Update the benchmark runner
- Switch the benchmark runner to use `m6id.2xlarge` as the benchmark.
This is for various reasons, but it mostly is a better equivalent for
the Polkadot Validator minimum.
- Benchmark runner had been a `c7i` for the last run, so the numbers
look like they increase more than they actually did.
- Ran all benchmarks

---------

Co-authored-by: Joe Caputo <joseph.caputo@projectliberty.io>
Co-authored-by: wilwade <wilwade@users.noreply.github.com>