Reduce Testnet WASM Size (#2355)
# Goal
The goal of this PR is to reduce the size of the released wasm for
Testnet

## Details

Testnet wasm has always been larger than mainnet, as it includes a few
more things like the sudo pallet. However in release
[v1.14.0](https://github.com/frequency-chain/frequency/releases/tag/v1.14.0)
it increased a large amount.

This has not been an issue, but in v1.16.0, the wasm increased to over
2.5Mb which caused issues with being able to upgrade cleanly on testnet.

A reason for the increase was that the mainnet genesis was also being
included in the testnet releases. This was not needed, so this PR
removes it from those builds.

## Summary

- Fixed release notes template issues
- Removed paseo local spec from the build for paseo testnet
- Limited the `frequency_genesis_config` to just `frequency` and
benchmarks runs
- Bumped the spec version

## Result

Local builds of the Mainnet wasm are at ~1.7M and Testnet dropped from
2.6M to 2.1M

## Local Testing

- Mainnet: `make build-mainnet-release && ls -lah
target/release/wbuild/frequency-runtime/frequency_runtime*`
- Testnet: `make build-testnet-release && ls -lah
target/release/wbuild/frequency-runtime/frequency_runtime*`