feat: custom runtime functions to generate deterministic Eth address from an MSA ID (#2358)
# Goal
This PR adds some internal pallet functions and custom runtime to
deterministically generate an Ethereum address from an MSA ID, and vice
versa.

The general approach is:
- Use a domain separator prefix to eliminate any possible conflict with
addresses generated by `CREATE2`
- Use a salt that is the hash of the string "MSA Generated"
- Take the last 20 bytes of the hash of `(prefix + MSA ID + salt)`
- If returning a string for the runtime query, pass through an [ERC-55
checksum
algorithm](https://github.com/ethereum/ercs/blob/master/ERCS/erc-55.md)
to alter the returned hex string

Note, although addresses generated in this manner cannot be reversed to
recover the related MSA ID, they can be verified (ie, given an address
and an MSA ID, we can determine whether or not the address is the result
of applying the above algorithm to the MSA ID). It is not envisioned
that there would ever be an _on-chain_ requirement to recover the MSA ID
from the address; if such a use case was desired for off-chain purposes,
it would be relatively trivial to populate a static lookup table mapping
all known or possible MSAs to their corresponding address, or
vice-versa.

Closes #2353 
Closes #2352 

# Discussion

- Q0: Currently returning ASCII bytes; or do we want to return the
binary address bytes?
- A: Now returning a struct containing both the binary address &
checksummed string

# Checklist
- [x] Updated Pallet Readme?
- [ ] Updated js/api-augment for Custom RPC APIs?
- [ ] Design doc(s) updated?
- [x] Unit Tests added?
- [x] e2e Tests added?
- [ ] Benchmarks added?
- [x] Spec version incremented?