Fix E2E Test Errors (#2405)
# Goal
The goal of this PR is to make the e2e tests work against Testnet on
Paseo

e2e Run against Testnet-Paseo:
https://github.com/frequency-chain/frequency/actions/runs/15022918786

All green except for non-released features!

# Discussion

The main issue is that many things need to wait until finalization for
the next thing to be able to consistently work.

- Updated JS packages
- Split up several e2e tests to parallelize
- Update the setup of many e2e tests to run in parallel
- Update the nonce setup to use Atomics so that it is "thread" safe
- Increase timeouts as the tests take longer against a chain that takes
time to finalize
- Suppress expected rpc errors on sendUnsigned