fix: add mechanism for configurable override of '--ignore-scripts' when running 'npm ci' (#2650)
# Goal
The goal of this PR is to enable selective running of npm install
scripts for packages in the dependency tree. This can be enabled in any
directory containing npm scripts (ie, 'package.json') simply by added a
file `package-exceptions.json` with the following structure:
```json
[
  {
    "package": "package-name",
    "script": "script-name"
  }
]
```
(NOTE, `script` is optional and defaults to "install" if not supplied)

The override behavior must be invoked by running the supplied script; it
is currently hooked in to building/running E2E tests in both the
Makefile script and CI pipeline.

Closes #2649