chore: create common action to setup Node.js and correct version of npm (#2652)
# Goal
The goal of this PR is to ensure that our CI pipeline is using at least
version `>= 11.5.1` of `npm` in order to support OID auth from GitHub.

It refactors many individual references to `actions/setup-node` into a
composite action that incorporates both `actions/setup-node` as well as
the specific commands to ensure a recent enough version of `npm`.

Another benefit is that the default version of Node.js for all of our CI
workflows is now specified in a single place (but can stlll be
overridden per invocation if necessary)